+++ src/Javacript/Helpers/Data/organizations.js
@@ -1,0 +1,28 @@
+const organizations = [
+  {
+    img: "images/orgImgs/oi_nss.png",
+    name: "nss-evening-cohort-14",
+    repos: 30,
+    topFive: ['productCards','petAdoption','sortingHat','gitSub','instaFam'],
+  },
+  {
+    img: "images/orgImgs/oi_org1.png",
+    name: "React Ladies",
+    repos: 32,
+    topFive: ['searchEngine','dataMapper','weatherApp','snowMap','gpsLocator'],
+  },
+  {
+    img: "images/orgImgs/oi_org2.png",
+    name: "TN Code Pros",
+    repos: 20,
+    topFive: ['trafficMap','liveMusicCalendar','restaurantRater','nhlStatsKeeper','barRaterApp'],
+  },
+  {
+    img: "images/orgImgs/oi_org3.png",
+    name: "Fortune 500 Devs",
+    repos: 27,
+    topFive: ['stockModel','facebookUserStats','tiktokAPI','blackRockFunds','appleOptimizer'],
+  },
+];
+
+export default organizations;
