+++ src/Javacript/Helpers/Data/projects.js
@@ -1,0 +1,29 @@
+const projects = [
+  {
+    name: 'my-goals',
+    description: 'No description',
+    date: new Date('2021-01-30T03:32:00'),
+  },
+  {
+    name: 'Team Project',
+    description: 'goals for first team project',
+    date: new Date('2020-12-01T08:00:00'),
+  },
+  {
+    name: 'Personal Project',
+    description: 'No description',
+    date: new Date('2021-02-02T12:30:00'),
+  },
+  {
+    name: 'Project 4',
+    description: 'No description',
+    date: new Date('2021-02-04T09:14:00'),
+  },
+  {
+    name: 'NSS-goals',
+    description: 'goals to achieve at NSS',
+    date: new Date('2021-02-02T10:00:00'),
+  },
+];
+
+export default projects;
