+++ src/Javacript/Helpers/Data/repos.js
@@ -1,0 +1,14 @@
+const repos = [
+  {
+    name: "example-repo",
+    description: "This is an example of what a repository will look like.",
+  },
+  {
+    name: "create-your-own-repo",
+    description: "Use the form below to create repositories of your own.",
+  },
+];
+
+const favoriteRepos = [];
+
+export { repos, favoriteRepos };
